+++ app/src/main/java/br/com/tagliaferro/soccerplayers/entities/LoginDTO.kt
@@ -1,0 +1,7 @@
+package br.com.tagliaferro.soccerplayers.entities
+
+data class LoginDTO(
+    val username: String,
+
+    val password: String
+)
