--- app/src/main/java/br/com/tagliaferro/soccerplayers/fragment/LoginFragment.kt
@@ -1,32 +1,0 @@
-package br.com.tagliaferro.soccerplayers.fragment
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Toast
-import androidx.fragment.app.Fragment
-import br.com.tagliaferro.soccerplayers.databinding.LoginFragmentBinding
-
-class LoginFragment : Fragment() {
-
-    private lateinit var binding: LoginFragmentBinding
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = LoginFragmentBinding.inflate(inflater, container, false)
-
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        binding.btnLogin.setOnClickListener {
-            Toast.makeText(view.context, "Teste with view binding", Toast.LENGTH_SHORT).show()
-        }
-    }
-}
